Migrate all but detail view and clinical item display
