+++ packages/terra-clinical-application/examples/test-examples/DefaultApplication.jsx
