+++ packages/terra-clinical-application/examples/test-examples/NoAppDelegateApplication.jsx
