+++ packages/terra-clinical-header/examples/test-examples/DefaultHeader.jsx
