+++ packages/terra-clinical-header/examples/test-examples/LeftAndRightContentHeader.jsx
