+++ packages/terra-clinical-header/examples/test-examples/LeftAndRightContentSubheader.jsx
