+++ packages/terra-clinical-header/examples/test-examples/LeftContentHeader.jsx
