+++ packages/terra-clinical-header/examples/test-examples/LongTitleAndContentHeader.jsx
