+++ packages/terra-clinical-header/examples/test-examples/LongTitleHeader.jsx
