+++ packages/terra-clinical-header/examples/test-examples/NoTitleRightContentHeader.jsx
