+++ packages/terra-clinical-header/examples/test-examples/RightContentHeader.jsx
