+++ packages/terra-clinical-item-collection/examples/test-examples/ItemCollectionAllElements.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ItemCollection from '../../../lib/ItemCollection';
+import ItemCollection from '../../lib/ItemCollection';
 
 const DefaultItemCollection = () => (
   <ItemCollection
