--- packages/terra-clinical-item-collection/tests/nightwatch/item-collection/ItemCollectionTestRoutes.jsx
@@ -1,26 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-
-import React from 'react';
-import { Route } from 'react-router';
-import ItemCollectionTests from './ItemCollectionTests';
-
-import ItemCollectionAllElements from './ItemCollectionAllElements';
-import ItemCollectionVaryingItems from './ItemCollectionVaryingItems';
-import ItemCollectionSelectable from './ItemCollectionSelectable';
-import TableStyles from './ItemCollectionTableStyles';
-import ListStyles from './ItemCollectionListStyles';
-import ItemStyles from './ItemCollectionItemStyles';
-
-const routes = (
-  <div>
-    <Route path="/tests/item-collection-tests" component={ItemCollectionTests} />
-    <Route path="/tests/item-collection-tests/all" component={ItemCollectionAllElements} />
-    <Route path="/tests/item-collection-tests/varying-items" component={ItemCollectionVaryingItems} />
-    <Route path="/tests/item-collection-tests/selectable" component={ItemCollectionSelectable} />
-    <Route path="/tests/item-collection-tests/table-styles" component={TableStyles} />
-    <Route path="/tests/item-collection-tests/list-styles" component={ListStyles} />
-    <Route path="/tests/item-collection-tests/item-styles" component={ItemStyles} />
-  </div>
-);
-
-export default routes;
