+++ packages/terra-clinical-item-view/examples/test-examples/AccessoryItemView.jsx
