+++ packages/terra-clinical-item-view/examples/test-examples/CommentItemView.jsx
