+++ packages/terra-clinical-item-view/examples/test-examples/DefaultItemView.jsx
