+++ packages/terra-clinical-item-view/examples/test-examples/DisplaysItemView.jsx
