+++ packages/terra-clinical-label-value-view/examples/test-examples/MultipleValueLabelValueView.jsx
