+++ packages/terra-clinical-label-value-view/examples/test-examples/TextValueLabelValueView.jsx
