+++ packages/terra-clinical-onset-picker/examples/test-examples/FiveYears.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import Base from 'terra-base';
 import moment from 'moment';
-import OnsetPicker from '../../../lib/OnsetPicker';
+import OnsetPicker from '../../lib/OnsetPicker';
 
 const locale = document.getElementsByTagName('html')[0].getAttribute('lang');
 
-export default class LessYear extends React.Component {
+class FiveYears extends React.Component {
   constructor(props) {
     super(props);
 
@@ -13,9 +13,9 @@ export default class LessYear extends React.Component {
     moment.now = () => new Date(2016, 8, 20);
 
     this.state = {
-      precision: OnsetPicker.Opts.Precisions.ONAT,
-      granularity: OnsetPicker.Opts.Granularities.MONTH,
-      onsetDate: moment(),
+      precision: OnsetPicker.Opts.Precisions.BEFORE,
+      granularity: OnsetPicker.Opts.Granularities.YEAR,
+      onsetDate: moment('2013-02-14'),
     };
 
     this.handleGranularity = this.handleGranularity.bind(this);
@@ -51,19 +51,22 @@ export default class LessYear extends React.Component {
   render() {
     return (
       <Base locale={locale}>
-        <p>Onset picker with minimum required fields and onChange handlers, supplied with a birthdate of 11 months old.</p>
-        <p>Current date locked to September 20, 2016</p>
-        <p>Testing age granularity selection restrictions.</p>
-
+        <p>Onset picker with onChange handlers, birthdate of 2011-09-20 (5 years).</p>
+        <p>Current date is locked to September 20, 2016</p>
+        <p>Testing returns from handlers are correct, if they are the json object should get updated on changes.</p>
+        <p>Also tests age caps for months and years.</p>
         <form onSubmit={this.handleSubmit}>
           <OnsetPicker
-            birthdate={moment().subtract(11, 'months').format()}
+            birthdate={moment('2011-09-20').format()}
+            granularity={this.state.granularity}
             granularitySelectName="test-granularity"
             granularitySelectOnChange={this.handleGranularity}
+            precision={this.state.precision}
             precisionSelectName="test-precision"
             precisionSelectOnChange={this.handlePrecision}
             onsetDateInputName="test-onsetDate"
             onsetDateInputOnChange={this.handleOnset}
+            onsetDate={moment(this.state.onsetDate).format()}
           />
           <button type="submit">Submit</button>
         </form>
@@ -72,3 +75,5 @@ export default class LessYear extends React.Component {
     );
   }
 }
+
+export default FiveYears;
