+++ packages/terra-clinical-onset-picker/tests/nightwatch/onset-spec.js
@@ -6,7 +6,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Displays a default Onset picker': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/default`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/default`)
       .assert.elementPresent('[data-terra-clinical-onset-picker="true"]')
       .assert.elementPresent('select[name="test-precision"]')
       .assert.elementPresent('select[name="test-granularity"]');
@@ -15,7 +15,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
   // Unknown precision
   'Displays no additional options when unknown precision selected': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/default`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/default`)
       .setValue('select[name="test-precision"]', 'unknown')
       .assert.elementNotPresent('select[name="test-granularity"]');
   },
@@ -23,7 +23,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
   // Year granularity
   'Displays year selection only when Year granularity selected': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/default`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/default`)
       .setValue('select[name="test-granularity"]', 'year')
       .assert.elementNotPresent('[data-terra-clinical-onset-picker="month"] select')
       .assert.elementPresent('[data-terra-clinical-onset-picker="year"] select');
@@ -31,7 +31,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Can select a year between the birthdate and current year': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/default`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/default`)
       .setValue('select[name="test-granularity"]', 'year')
       .setValue('[data-terra-clinical-onset-picker="year"] select', 2015)
       .expect.element('[data-terra-clinical-onset-picker="year"] select').to.have.value.equals(2015);
@@ -39,7 +39,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Cannot select a year before the birthdate': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/default`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/default`)
       .setValue('select[name="test-granularity"]', 'year')
       .setValue('[data-terra-clinical-onset-picker="year"] select', 2010)
       .expect.element('[data-terra-clinical-onset-picker="year"] select').to.have.value.not.equals(2010);
@@ -47,7 +47,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Cannot select a year in the future': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/default`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/default`)
       .setValue('select[name="test-granularity"]', 'year')
       .setValue('[data-terra-clinical-onset-picker="year"] select', 2017)
       .expect.element('[data-terra-clinical-onset-picker="year"] select').to.have.value.not.equals(2017);
@@ -56,14 +56,14 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
   // Month Granularity
   'Displays month and year selection when Month/Year granularity selected': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/default`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/default`)
       .setValue('select[name="test-granularity"]', 'month')
       .assert.elementPresent('[data-terra-clinical-onset-picker="month"] select')
       .assert.elementPresent('[data-terra-clinical-onset-picker="year"] select');
   },
 
   'Can select month between birthdate and current date': (browser) => {
-    browser.url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/default`);
+    browser.url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/default`);
 
     browser.setValue('select[name="test-granularity"]', 'month')
       .setValue('[data-terra-clinical-onset-picker="year"] select', 2015)
@@ -74,7 +74,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Cannot select a month before the birthdate': (browser) => {
     browser
-      .url(`${browser.launcherUrl}/#/raw/tests/clinical-onset-picker/component/default`)
+      .url(`${browser.launcherUrl}/#/raw/tests/clinical-onset-picker/default`)
       .setValue('select[name="test-granularity"]', 'month')
       .setValue('[data-terra-clinical-onset-picker="year"] select', 2011)
       .setValue('[data-terra-clinical-onset-picker="month"] select', 'February')
@@ -83,7 +83,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Cannot select a month in the future': (browser) => {
     browser
-      .url(`${browser.launcherUrl}/#/raw/tests/clinical-onset-picker/component/default`)
+      .url(`${browser.launcherUrl}/#/raw/tests/clinical-onset-picker/default`)
       .setValue('select[name="test-granularity"]', 'month')
       .setValue('[data-terra-clinical-onset-picker="year"] select', 2016)
       .setValue('[data-terra-clinical-onset-picker="month"] select', 'October')
@@ -93,7 +93,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
   // Date granularity
   'Displays date input when Date granularity is selected': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/default`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/default`)
       .setValue('select[name="test-granularity"]', 'date')
       .assert.elementNotPresent('[data-terra-clinical-onset-picker="year"] select')
       .assert.elementNotPresent('[data-terra-clinical-onset-picker="month"] select')
@@ -103,7 +103,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
   // Age granularity
   'Displays age inputs when Age granularity is selected': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/default`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/default`)
       .setValue('select[name="test-granularity"]', 'age')
       .assert.elementPresent('[data-terra-clinical-onset-picker="age"] input')
       .assert.elementPresent('[data-terra-clinical-onset-picker="age_unit"] select');
@@ -111,7 +111,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Cannot select number of years greater than age': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/five-years`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/five-years`)
       .setValue('select[name="test-granularity"]', 'age')
       .setValue('[data-terra-clinical-onset-picker="age_unit"] select', 'years')
       .clearValue('[data-terra-clinical-onset-picker="age"] input')
@@ -122,7 +122,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Cannot select more than 24 months': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/five-years`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/five-years`)
       .setValue('select[name="test-granularity"]', 'age')
       .setValue('[data-terra-clinical-onset-picker="age_unit"] select', 'months')
       .clearValue('[data-terra-clinical-onset-picker="age"] input')
@@ -133,7 +133,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Cannot select number of months that exceede age': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/less-year`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/less-year`)
       .setValue('select[name="test-granularity"]', 'age')
       .setValue('[data-terra-clinical-onset-picker="age_unit"] select', 'months')
       .clearValue('[data-terra-clinical-onset-picker="age"] input')
@@ -144,7 +144,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Cannot select more than 8 weeks': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/five-years`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/five-years`)
       .setValue('select[name="test-granularity"]', 'age')
       .setValue('[data-terra-clinical-onset-picker="age_unit"] select', 'weeks')
       .clearValue('[data-terra-clinical-onset-picker="age"] input')
@@ -155,7 +155,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Cannot select number of weeks that exceede age': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/less-month`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/less-month`)
       .setValue('select[name="test-granularity"]', 'age')
       .setValue('[data-terra-clinical-onset-picker="age_unit"] select', 'weeks')
       .clearValue('[data-terra-clinical-onset-picker="age"] input')
@@ -166,7 +166,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Cannot select years duration if age is less than a year old': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/less-year`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/less-year`)
       .setValue('select[name="test-granularity"]', 'age')
       .setValue('[data-terra-clinical-onset-picker="age_unit"] select', 'years')
       .expect.element('[data-terra-clinical-onset-picker="age_unit"]').to.have.value.not.equals('years');
@@ -174,7 +174,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'Cannot select months duration if age is less than a month old': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/less-month`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/less-month`)
       .setValue('select[name="test-granularity"]', 'age')
       .setValue('[data-terra-clinical-onset-picker="age_unit"] select', 'months')
       .expect.element('[data-terra-clinical-onset-picker="age_unit"]').to.have.value.not.equals('months');
@@ -183,7 +183,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
   // onChange handlers
   'When precision is changed an event is fired to the precisionSelectOnChange callback': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/five-years`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/five-years`)
       .setValue('select[name="test-precision"]', 'unknown')
       .click('button[type="submit"]')
       .expect.element('[data-test-json="true"]').text.to.contain('unknown');
@@ -191,7 +191,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'When granularity is changed an event is fired to the granularitySelectOnChange callback': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/five-years`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/five-years`)
       .setValue('select[name="test-granularity"]', 'age')
       .click('button[type="submit"]')
       .expect.element('[data-test-json="true"]').text.to.contain('age');
@@ -199,7 +199,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'When onsetDate is changed by the month select an event is fired to the onsetDateInputOnChange callback': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/five-years`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/five-years`)
       .setValue('select[name="test-granularity"]', 'month')
       .setValue('[data-terra-clinical-onset-picker="month"] select', 'October')
       .click('button[type="submit"]')
@@ -208,7 +208,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'When onsetDate is changed by the year select an event is fired to the onsetDateInputOnChange callback': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/five-years`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/five-years`)
       .setValue('select[name="test-granularity"]', 'year')
       .setValue('[data-terra-clinical-onset-picker="year"] select', 2016)
       .click('button[type="submit"]')
@@ -217,7 +217,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'When onsetDate is changed by the age input an event is fired to the onsetDateInputOnChange callback': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/five-years`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/five-years`)
       .setValue('select[name="test-granularity"]', 'age')
       .setValue('[data-terra-clinical-onset-picker="age_unit"] select', 'years')
       .clearValue('[data-terra-clinical-onset-picker="age"] input') // CLEAR FIRST, SET VALUE APPENDS NEW VALUE
@@ -228,7 +228,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
 
   'When onsetDate is changed by the ageUnit select an event is fired to the onsetDateInputOnChange callback': (browser) => {
     browser
-      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/component/five-years`)
+      .url(`${browser.launchUrl}/#/raw/tests/clinical-onset-picker/five-years`)
       .setValue('select[name="test-granularity"]', 'age')
       .clearValue('[data-terra-clinical-onset-picker="age"] input') // CLEAR FIRST, SET VALUE APPENDS NEW VALUE
       .setValue('[data-terra-clinical-onset-picker="age"] input', 4)
